Reuse RWRecorder for http and external http channels.
